refactor: 도착지점까지의 경로 구하는 메소드 이동
