+++ src/test/java/chess/domain/piece/KingTest.java
@@ -39,15 +39,4 @@ class KingTest {
                 .doesNotThrowAnyException();
 
     }
-
-    @Test
-    @DisplayName("목적지까지 경로를 구한다.")
-    void findRoute() {
-        // given
-        ChessPiece king = new King(Color.BLACK);
-        // when
-        Stack<Position> actual = king.findRoute(initialPosition, new Position("e4"));
-        // then
-        assertThat(actual).isEmpty();
-    }
 }
