+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -40,14 +40,4 @@ class KnightTest {
 
     }
 
-    @Test
-    @DisplayName("목적지까지 경로를 구한다.")
-    void findRoute() {
-        // given
-        ChessPiece knight = new Knight(Color.BLACK);
-        // when
-        Stack<Position> actual = knight.findRoute(initialPosition, new Position("b6"));
-        // then
-        assertThat(actual).isEmpty();
-    }
 }
